Attempt to correct a recurring mapping bug
The AniDB - TVDB third party mapping files are sometimes a bit crude around long running shows and simply map them as 'absolute', this unfortunately doesn't work with many of their specials. Whilst this can be fixed using the correction file or submitting a fix to the mapping source I've added a few lines of code to try and mitigate this mildly frequent occurrence.