Stripped out persistent path helper
It was a good idea but isn't practical due to the complete disconnect
between seatch and update.

*The idea was to have 'fix incorrect match' changes be persistent, so if
additional media was added later then the 'fix' would be automatically
applied; it would have had to be based both on series name or folder but
manual searches overwite the name and getting the media folder from the
update function isn't consistent as you've dealing with a list of files*

Will revisit this in the future.